+++ src/app/mod.rs
@@ -1,3 +1,4 @@
 pub mod ddns;
 pub mod playlists;
 pub mod thumbnails;
+pub mod vfs;
